Fixed issue not finding root for mono
Using Windows Server 2008 R2 Datacenter and Mono 4.6.1 the debugger
would fail when starting debugging. The fail was caused by not been able
to find the pdb2mdb.bat file to convert the pdb files. THe reason been
it was looking in the wrong place int he registry
software/novell/mono/{version} but mono now seems to install with just
software/mono and no version number.

The fix i have submitted preserves the novell registry search as i
presume this is for older mono installs, but works with the new registry
entries.